Update test evidence files with newer ones that has cpu family, id and stepping.
This is to enable testing the dynamic policy manager.
I also added evidence from both Genoa and Milan that runs the same exact stage0.
However, I had to modify the images to extract the application signing keys.

The current evidence do not have cpu ids set (I have done a similar update to Oak's
https://oak-review.git.corp.google.com/c/oak/+/26100 ).

Although the changes are only to unit-tests, it seems that any change to `mod tests`,
change the binary digests, and for that I updated the policies.

Bug: b/434016988
WANT_LGMT=any
Change-Id: Iec0ae9b378a2c3c5cf43f805f65e6b2d425bcb61