Fetch Stage 0 blob data for dynamic attestation GCP
Notes: To support dynamic attestation, the verifier requires access to
Stage 0 binary blobs that are associated with appraisal policies. This
change updates the GCP policy fetcher to retrieve this data from
Spanner.

A `DYNAMIC_ATTESTATION` compile-time definition is introduced to
conditionally alter the fetching logic. When enabled, the SQL queries in
`GetLatestNPolicies` and `GetLatestNPoliciesForDigest` are modified to
join with the `Stage0BlobToPolicy` and `Stage0Blobs` tables.

The retrieved blob data is populated into the
`stage0_binary_sha256_to_blob` map within the `AppraisalPolicies` proto.
A helper map is used during processing to ensure that if a policy is
associated with multiple Stage 0 blobs, it is not duplicated in the
final result.

Test: bazel test //tvs/appraisal_policies:policy-fetcher-gcp_test
--config=dynamic_attestation

Bug: b/434016988
Change-Id: Ie545cbf0156140d4d1b6f7c79caa2139c4089ea4