Adapt C++ tests to support dynamic attestation
Update the C++ tests for the launcher server and TVS service to run
correctly when the dynamic attestation feature is enabled. This allows
for testing both the static policy mechanism and the new dynamic policy
fetching within the same test suite.

The tests now use conditional compilation to select the appropriate
attestation policies(dynamic or static). The BUILD files are also
updated to add the test_util_cc helper (policy_generator) when the
feature is enabled.

New test utilities are also added to `policy_generator` to facilitate
more complex dynamic policy tests(TvsService.DynamicPolicyFetching):
* `CreateMixedDynamicAndInsecurePolicies`: Generates a policy set
  with both a dynamic and an insecure policy to test selection logic.
* `PopulateTempBlobDirectory`: Sets up the stage0 blob directory required for the policy fetcher to locate blobs for dynamic policies.

This is the final change for all tests to pass under the
dynamic_attestation flag.

Test: bazel test //... --config=dynamic_attestation

Bug: b/434016988
Change-Id: Ieecb1d4fc950135889a398702b1fe2b557a5e288