Allow all vCPUs from 1-256 to be accepted by all policies
Notes:
Instead of keeping an allowlist of vCPU counts in the appraisal
policies, we want to loop through all vCPU counts from 1-256 as
acceptable vCPU counts. This allows any vCPU count to be accepted,
meaning that support for a new CVM configuration with a change in vCPU
can be supported without additional signing and rollout of an updated
appraisal policy.

Although this change introduces increased computational load, it is
aligned with the goal to allow support for any vCPU without needing
appraisal policy revision. In the future, having the CVMs vCPU passed in
through the attestation report will eliminate this computational load
completely.

Bug: b/434016988
Change-Id: I950cb23a03315806d8be7a842fc11954f69eae8b